requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.0
pdfplumber
easyocr
pdf2image
transformers
psutil
torch
sentencepiece
sacremoses
